[docs] Improve custom field input demos readability
