+++ docs/data/date-pickers/custom-field/PickerWithBrowserField.js
@@ -1,18 +1,9 @@
 import * as React from 'react';
 
 import { unstable_useForkRef as useForkRef } from '@mui/utils';
-import { useSlotProps } from '@mui/base/utils';
 import Box from '@mui/material/Box';
-import Stack from '@mui/material/Stack';
-import IconButton from '@mui/material/IconButton';
-import InputAdornment from '@mui/material/InputAdornment';
-import { DemoContainer } from '@mui/x-date-pickers/internals/demo';
-import { DateRangeIcon } from '@mui/x-date-pickers/icons';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { DateRangePicker } from '@mui/x-date-pickers-pro/DateRangePicker';
-import { unstable_useSingleInputDateRangeField as useSingleInputDateRangeField } from '@mui/x-date-pickers-pro/SingleInputDateRangeField';
-import { unstable_useMultiInputDateRangeField as useMultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDateRangeField';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { unstable_useDateField as useDateField } from '@mui/x-date-pickers/DateField';
 
@@ -37,7 +28,7 @@ const BrowserField = React.forwardRef((props, ref) => {
 
   return (
     <Box
-      sx={{ ...(sx || {}), display: 'flex', alignItems: 'center', flexGrow: 1 }}
+      sx={{ ...(sx || {}), display: 'flex', alignItems: 'center' }}
       id={id}
       ref={handleRef}
     >
@@ -45,166 +36,6 @@ const BrowserField = React.forwardRef((props, ref) => {
       <input disabled={disabled} ref={inputRef} {...other} />
       {endAdornment}
     </Box>
-  );
-});
-
-const BrowserSingleInputDateRangeField = React.forwardRef((props, ref) => {
-  const { slots, slotProps, onAdornmentClick, ...other } = props;
-
-  const { inputRef: externalInputRef, ...textFieldProps } = useSlotProps({
-    elementType: 'input',
-    externalSlotProps: slotProps?.textField,
-    externalForwardedProps: other,
-    ownerState: props,
-  });
-
-  const {
-    ref: inputRef,
-    onClear,
-    clearable,
-    ...fieldProps
-  } = useSingleInputDateRangeField({
-    props: textFieldProps,
-    inputRef: externalInputRef,
-  });
-
-  /* If you don't need a clear button, you can skip the use of this hook */
-  const { InputProps: ProcessedInputProps, fieldProps: processedFieldProps } =
-    useClearableField({
-      onClear,
-      clearable,
-      fieldProps,
-      InputProps: {
-        ...fieldProps.InputProps,
-        endAdornment: (
-          <InputAdornment position="end">
-            <IconButton onClick={onAdornmentClick}>
-              <DateRangeIcon />
-            </IconButton>
-          </InputAdornment>
-        ),
-      },
-      slots,
-      slotProps,
-    });
-
-  return (
-    <BrowserField
-      {...processedFieldProps}
-      ref={ref}
-      style={{
-        minWidth: 300,
-      }}
-      inputRef={inputRef}
-      InputProps={{ ...ProcessedInputProps }}
-    />
-  );
-});
-
-BrowserSingleInputDateRangeField.fieldType = 'single-input';
-
-const BrowserSingleInputDateRangePicker = React.forwardRef((props, ref) => {
-  const [isOpen, setIsOpen] = React.useState(false);
-
-  const toggleOpen = () => setIsOpen((currentOpen) => !currentOpen);
-
-  const handleOpen = () => setIsOpen(true);
-
-  const handleClose = () => setIsOpen(false);
-
-  return (
-    <DateRangePicker
-      ref={ref}
-      {...props}
-      open={isOpen}
-      onClose={handleClose}
-      onOpen={handleOpen}
-      slots={{ field: BrowserSingleInputDateRangeField }}
-      slotProps={{
-        ...props?.slotProps,
-        field: {
-          onAdornmentClick: toggleOpen,
-          ...props?.slotProps?.field,
-        },
-      }}
-    />
-  );
-});
-
-const BrowserMultiInputDateRangeField = React.forwardRef((props, ref) => {
-  const {
-    slotProps,
-    value,
-    defaultValue,
-    format,
-    onChange,
-    readOnly,
-    disabled,
-    onError,
-    shouldDisableDate,
-    minDate,
-    maxDate,
-    disableFuture,
-    disablePast,
-    selectedSections,
-    onSelectedSectionsChange,
-    className,
-  } = props;
-
-  const { inputRef: startInputRef, ...startTextFieldProps } = useSlotProps({
-    elementType: 'input',
-    externalSlotProps: slotProps?.textField,
-    ownerState: { ...props, position: 'start' },
-  });
-
-  const { inputRef: endInputRef, ...endTextFieldProps } = useSlotProps({
-    elementType: 'input',
-    externalSlotProps: slotProps?.textField,
-    ownerState: { ...props, position: 'end' },
-  });
-
-  const {
-    startDate: { ref: startRef, ...startDateProps },
-    endDate: { ref: endRef, ...endDateProps },
-  } = useMultiInputDateRangeField({
-    sharedProps: {
-      value,
-      defaultValue,
-      format,
-      onChange,
-      readOnly,
-      disabled,
-      onError,
-      shouldDisableDate,
-      minDate,
-      maxDate,
-      disableFuture,
-      disablePast,
-      selectedSections,
-      onSelectedSectionsChange,
-    },
-    startTextFieldProps,
-    endTextFieldProps,
-    startInputRef,
-    endInputRef,
-  });
-
-  return (
-    <Stack ref={ref} spacing={2} direction="row" className={className}>
-      <BrowserField {...startDateProps} inputRef={startRef} />
-      <span> — </span>
-      <BrowserField {...endDateProps} inputRef={endRef} />
-    </Stack>
-  );
-});
-
-const BrowserDateRangePicker = React.forwardRef((props, ref) => {
-  return (
-    <DateRangePicker
-      ref={ref}
-      {...props}
-      slots={{ ...props?.slots, field: BrowserMultiInputDateRangeField }}
-    />
   );
 });
 
@@ -254,21 +85,11 @@ const BrowserDatePicker = React.forwardRef((props, ref) => {
 export default function PickerWithBrowserField() {
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
-      <DemoContainer
-        components={['DatePicker', 'SingleInputDateRangeField', 'DateRangePicker']}
-      >
-        <BrowserDatePicker
-          slotProps={{
-            field: { clearable: true },
-          }}
-        />
-        <BrowserSingleInputDateRangePicker
-          slotProps={{
-            field: { clearable: true },
-          }}
-        />
-        <BrowserDateRangePicker />
-      </DemoContainer>
+      <BrowserDatePicker
+        slotProps={{
+          field: { clearable: true },
+        }}
+      />
     </LocalizationProvider>
   );
 }
