+++ docs/data/date-pickers/custom-field/PickerWithBrowserField.tsx
@@ -1,24 +1,9 @@
 import * as React from 'react';
 import { Dayjs } from 'dayjs';
 import { unstable_useForkRef as useForkRef } from '@mui/utils';
-import { useSlotProps } from '@mui/base/utils';
 import Box from '@mui/material/Box';
-import Stack from '@mui/material/Stack';
-import IconButton from '@mui/material/IconButton';
-import InputAdornment from '@mui/material/InputAdornment';
-import { DemoContainer } from '@mui/x-date-pickers/internals/demo';
-import { DateRangeIcon } from '@mui/x-date-pickers/icons';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import {
-  DateRangePicker,
-  DateRangePickerProps,
-} from '@mui/x-date-pickers-pro/DateRangePicker';
-import {
-  unstable_useSingleInputDateRangeField as useSingleInputDateRangeField,
-  SingleInputDateRangeFieldProps,
-} from '@mui/x-date-pickers-pro/SingleInputDateRangeField';
-import { unstable_useMultiInputDateRangeField as useMultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDateRangeField';
 import { DatePicker, DatePickerProps } from '@mui/x-date-pickers/DatePicker';
 import {
   unstable_useDateField as useDateField,
@@ -30,20 +15,10 @@ import {
 } from '@mui/x-date-pickers/DateField/DateField.types';
 import { useClearableField } from '@mui/x-date-pickers/hooks';
 import {
-  BaseMultiInputFieldProps,
-  DateRange,
-  DateRangeValidationError,
-  UseDateRangeFieldProps,
-  MultiInputFieldSlotTextFieldProps,
   BaseSingleInputFieldProps,
   DateValidationError,
-  RangeFieldSection,
   FieldSection,
 } from '@mui/x-date-pickers-pro';
-import type {
-  SingleInputDateRangeFieldSlotsComponent,
-  SingleInputDateRangeFieldSlotsComponentsProps,
-} from '@mui/x-date-pickers-pro/SingleInputDateRangeField/SingleInputDateRangeField.types';
 
 interface BrowserFieldProps
   extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'> {
@@ -84,7 +59,7 @@ const BrowserField = React.forwardRef(
 
     return (
       <Box
-        sx={{ ...(sx || {}), display: 'flex', alignItems: 'center', flexGrow: 1 }}
+        sx={{ ...(sx || {}), display: 'flex', alignItems: 'center' }}
         id={id}
         ref={handleRef}
       >
@@ -95,207 +70,6 @@ const BrowserField = React.forwardRef(
     );
   },
 ) as BrowserFieldComponent;
-
-interface BrowserSingleInputDateRangeFieldProps
-  extends SingleInputDateRangeFieldProps<
-    Dayjs,
-    Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'>
-  > {
-  onAdornmentClick?: () => void;
-}
-
-type BrowserSingleInputDateRangeFieldComponent = ((
-  props: BrowserSingleInputDateRangeFieldProps & React.RefAttributes<HTMLDivElement>,
-) => React.JSX.Element) & { fieldType?: string };
-
-const BrowserSingleInputDateRangeField = React.forwardRef(
-  (props: BrowserSingleInputDateRangeFieldProps, ref: React.Ref<HTMLDivElement>) => {
-    const { slots, slotProps, onAdornmentClick, ...other } = props;
-
-    const {
-      inputRef: externalInputRef,
-      ...textFieldProps
-    }: SingleInputDateRangeFieldProps<Dayjs> = useSlotProps({
-      elementType: 'input',
-      externalSlotProps: slotProps?.textField,
-      externalForwardedProps: other,
-      ownerState: props as any,
-    });
-
-    const {
-      ref: inputRef,
-      onClear,
-      clearable,
-      ...fieldProps
-    } = useSingleInputDateRangeField<Dayjs, typeof textFieldProps>({
-      props: textFieldProps,
-      inputRef: externalInputRef,
-    });
-
-    /* If you don't need a clear button, you can skip the use of this hook */
-    const { InputProps: ProcessedInputProps, fieldProps: processedFieldProps } =
-      useClearableField<
-        {},
-        typeof textFieldProps.InputProps,
-        SingleInputDateRangeFieldSlotsComponent,
-        SingleInputDateRangeFieldSlotsComponentsProps<Dayjs>
-      >({
-        onClear,
-        clearable,
-        fieldProps,
-        InputProps: {
-          ...fieldProps.InputProps,
-          endAdornment: (
-            <InputAdornment position="end">
-              <IconButton onClick={onAdornmentClick}>
-                <DateRangeIcon />
-              </IconButton>
-            </InputAdornment>
-          ),
-        },
-        slots,
-        slotProps,
-      });
-
-    return (
-      <BrowserField
-        {...processedFieldProps}
-        ref={ref}
-        style={{
-          minWidth: 300,
-        }}
-        inputRef={inputRef}
-        InputProps={{ ...ProcessedInputProps }}
-      />
-    );
-  },
-) as BrowserSingleInputDateRangeFieldComponent;
-
-BrowserSingleInputDateRangeField.fieldType = 'single-input';
-
-const BrowserSingleInputDateRangePicker = React.forwardRef(
-  (props: DateRangePickerProps<Dayjs>, ref: React.Ref<HTMLDivElement>) => {
-    const [isOpen, setIsOpen] = React.useState(false);
-
-    const toggleOpen = () => setIsOpen((currentOpen) => !currentOpen);
-
-    const handleOpen = () => setIsOpen(true);
-
-    const handleClose = () => setIsOpen(false);
-
-    return (
-      <DateRangePicker
-        ref={ref}
-        {...props}
-        open={isOpen}
-        onClose={handleClose}
-        onOpen={handleOpen}
-        slots={{ field: BrowserSingleInputDateRangeField }}
-        slotProps={{
-          ...props?.slotProps,
-          field: {
-            onAdornmentClick: toggleOpen,
-            ...props?.slotProps?.field,
-          } as any,
-        }}
-      />
-    );
-  },
-);
-
-interface BrowserMultiInputDateRangeFieldProps
-  extends UseDateRangeFieldProps<Dayjs>,
-    BaseMultiInputFieldProps<
-      DateRange<Dayjs>,
-      Dayjs,
-      RangeFieldSection,
-      DateRangeValidationError
-    > {}
-
-type BrowserMultiInputDateRangeFieldComponent = ((
-  props: BrowserMultiInputDateRangeFieldProps & React.RefAttributes<HTMLDivElement>,
-) => React.JSX.Element) & { propTypes?: any };
-
-const BrowserMultiInputDateRangeField = React.forwardRef(
-  (props: BrowserMultiInputDateRangeFieldProps, ref: React.Ref<HTMLDivElement>) => {
-    const {
-      slotProps,
-      value,
-      defaultValue,
-      format,
-      onChange,
-      readOnly,
-      disabled,
-      onError,
-      shouldDisableDate,
-      minDate,
-      maxDate,
-      disableFuture,
-      disablePast,
-      selectedSections,
-      onSelectedSectionsChange,
-      className,
-    } = props;
-
-    const { inputRef: startInputRef, ...startTextFieldProps } = useSlotProps({
-      elementType: 'input',
-      externalSlotProps: slotProps?.textField,
-      ownerState: { ...props, position: 'start' },
-    }) as MultiInputFieldSlotTextFieldProps;
-
-    const { inputRef: endInputRef, ...endTextFieldProps } = useSlotProps({
-      elementType: 'input',
-      externalSlotProps: slotProps?.textField,
-      ownerState: { ...props, position: 'end' },
-    }) as MultiInputFieldSlotTextFieldProps;
-
-    const {
-      startDate: { ref: startRef, ...startDateProps },
-      endDate: { ref: endRef, ...endDateProps },
-    } = useMultiInputDateRangeField<Dayjs, MultiInputFieldSlotTextFieldProps>({
-      sharedProps: {
-        value,
-        defaultValue,
-        format,
-        onChange,
-        readOnly,
-        disabled,
-        onError,
-        shouldDisableDate,
-        minDate,
-        maxDate,
-        disableFuture,
-        disablePast,
-        selectedSections,
-        onSelectedSectionsChange,
-      },
-      startTextFieldProps,
-      endTextFieldProps,
-      startInputRef,
-      endInputRef,
-    });
-
-    return (
-      <Stack ref={ref} spacing={2} direction="row" className={className}>
-        <BrowserField {...startDateProps} inputRef={startRef} />
-        <span> — </span>
-        <BrowserField {...endDateProps} inputRef={endRef} />
-      </Stack>
-    );
-  },
-) as BrowserMultiInputDateRangeFieldComponent;
-
-const BrowserDateRangePicker = React.forwardRef(
-  (props: DateRangePickerProps<Dayjs>, ref: React.Ref<HTMLDivElement>) => {
-    return (
-      <DateRangePicker
-        ref={ref}
-        {...props}
-        slots={{ ...props?.slots, field: BrowserMultiInputDateRangeField }}
-      />
-    );
-  },
-);
 
 interface BrowserDateFieldProps
   extends UseDateFieldProps<Dayjs>,
@@ -366,21 +140,11 @@ const BrowserDatePicker = React.forwardRef(
 export default function PickerWithBrowserField() {
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
-      <DemoContainer
-        components={['DatePicker', 'SingleInputDateRangeField', 'DateRangePicker']}
-      >
-        <BrowserDatePicker
-          slotProps={{
-            field: { clearable: true },
-          }}
-        />
-        <BrowserSingleInputDateRangePicker
-          slotProps={{
-            field: { clearable: true },
-          }}
-        />
-        <BrowserDateRangePicker />
-      </DemoContainer>
+      <BrowserDatePicker
+        slotProps={{
+          field: { clearable: true },
+        }}
+      />
     </LocalizationProvider>
   );
 }
