+++ docs/data/date-pickers/custom-field/RangePickerWithBrowserField.js
@@ -1,0 +1,125 @@
+import * as React from 'react';
+
+import { unstable_useForkRef as useForkRef } from '@mui/utils';
+import { useSlotProps } from '@mui/base/utils';
+import Box from '@mui/material/Box';
+import Stack from '@mui/material/Stack';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import { DateRangePicker } from '@mui/x-date-pickers-pro/DateRangePicker';
+import { unstable_useMultiInputDateRangeField as useMultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDateRangeField';
+
+const BrowserField = React.forwardRef((props, ref) => {
+  const {
+    disabled,
+    id,
+    label,
+    inputRef,
+    InputProps: { ref: containerRef, startAdornment, endAdornment } = {},
+    // extracting `error`, 'focused', and `ownerState` as `input` does not support those props
+    error,
+    focused,
+    ownerState,
+    sx,
+    ...other
+  } = props;
+
+  const handleRef = useForkRef(containerRef, ref);
+
+  return (
+    <Box
+      sx={{ ...(sx || {}), display: 'flex', alignItems: 'center', flexGrow: 1 }}
+      id={id}
+      ref={handleRef}
+    >
+      {startAdornment}
+      <input disabled={disabled} ref={inputRef} {...other} />
+      {endAdornment}
+    </Box>
+  );
+});
+
+const BrowserMultiInputDateRangeField = React.forwardRef((props, ref) => {
+  const {
+    slotProps,
+    value,
+    defaultValue,
+    format,
+    onChange,
+    readOnly,
+    disabled,
+    onError,
+    shouldDisableDate,
+    minDate,
+    maxDate,
+    disableFuture,
+    disablePast,
+    selectedSections,
+    onSelectedSectionsChange,
+    className,
+  } = props;
+
+  const { inputRef: startInputRef, ...startTextFieldProps } = useSlotProps({
+    elementType: 'input',
+    externalSlotProps: slotProps?.textField,
+    ownerState: { ...props, position: 'start' },
+  });
+
+  const { inputRef: endInputRef, ...endTextFieldProps } = useSlotProps({
+    elementType: 'input',
+    externalSlotProps: slotProps?.textField,
+    ownerState: { ...props, position: 'end' },
+  });
+
+  const {
+    startDate: { ref: startRef, ...startDateProps },
+    endDate: { ref: endRef, ...endDateProps },
+  } = useMultiInputDateRangeField({
+    sharedProps: {
+      value,
+      defaultValue,
+      format,
+      onChange,
+      readOnly,
+      disabled,
+      onError,
+      shouldDisableDate,
+      minDate,
+      maxDate,
+      disableFuture,
+      disablePast,
+      selectedSections,
+      onSelectedSectionsChange,
+    },
+    startTextFieldProps,
+    endTextFieldProps,
+    startInputRef,
+    endInputRef,
+  });
+
+  return (
+    <Stack ref={ref} spacing={2} direction="row" className={className}>
+      <BrowserField {...startDateProps} inputRef={startRef} />
+      <span> — </span>
+      <BrowserField {...endDateProps} inputRef={endRef} />
+    </Stack>
+  );
+});
+
+const BrowserDateRangePicker = React.forwardRef((props, ref) => {
+  return (
+    <DateRangePicker
+      ref={ref}
+      {...props}
+      slots={{ ...props?.slots, field: BrowserMultiInputDateRangeField }}
+    />
+  );
+});
+
+export default function RangePickerWithBrowserField() {
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <BrowserDateRangePicker />
+    </LocalizationProvider>
+  );
+}
