+++ docs/data/date-pickers/custom-field/RangePickerWithBrowserField.tsx
@@ -1,0 +1,174 @@
+import * as React from 'react';
+import { Dayjs } from 'dayjs';
+import { unstable_useForkRef as useForkRef } from '@mui/utils';
+import { useSlotProps } from '@mui/base/utils';
+import Box from '@mui/material/Box';
+import Stack from '@mui/material/Stack';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import {
+  DateRangePicker,
+  DateRangePickerProps,
+} from '@mui/x-date-pickers-pro/DateRangePicker';
+import { unstable_useMultiInputDateRangeField as useMultiInputDateRangeField } from '@mui/x-date-pickers-pro/MultiInputDateRangeField';
+import {
+  BaseMultiInputFieldProps,
+  DateRange,
+  DateRangeValidationError,
+  UseDateRangeFieldProps,
+  MultiInputFieldSlotTextFieldProps,
+  RangeFieldSection,
+} from '@mui/x-date-pickers-pro';
+
+interface BrowserFieldProps
+  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'> {
+  label?: React.ReactNode;
+  inputRef?: React.Ref<any>;
+  InputProps?: {
+    ref?: React.Ref<any>;
+    endAdornment?: React.ReactNode;
+    startAdornment?: React.ReactNode;
+  };
+  error?: boolean;
+  focused?: boolean;
+  ownerState?: any;
+  sx?: any;
+}
+
+type BrowserFieldComponent = ((
+  props: BrowserFieldProps & React.RefAttributes<HTMLDivElement>,
+) => React.JSX.Element) & { propTypes?: any };
+
+const BrowserField = React.forwardRef(
+  (props: BrowserFieldProps, ref: React.Ref<HTMLDivElement>) => {
+    const {
+      disabled,
+      id,
+      label,
+      inputRef,
+      InputProps: { ref: containerRef, startAdornment, endAdornment } = {},
+      // extracting `error`, 'focused', and `ownerState` as `input` does not support those props
+      error,
+      focused,
+      ownerState,
+      sx,
+      ...other
+    } = props;
+
+    const handleRef = useForkRef(containerRef, ref);
+
+    return (
+      <Box
+        sx={{ ...(sx || {}), display: 'flex', alignItems: 'center', flexGrow: 1 }}
+        id={id}
+        ref={handleRef}
+      >
+        {startAdornment}
+        <input disabled={disabled} ref={inputRef} {...other} />
+        {endAdornment}
+      </Box>
+    );
+  },
+) as BrowserFieldComponent;
+
+interface BrowserMultiInputDateRangeFieldProps
+  extends UseDateRangeFieldProps<Dayjs>,
+    BaseMultiInputFieldProps<
+      DateRange<Dayjs>,
+      Dayjs,
+      RangeFieldSection,
+      DateRangeValidationError
+    > {}
+
+type BrowserMultiInputDateRangeFieldComponent = ((
+  props: BrowserMultiInputDateRangeFieldProps & React.RefAttributes<HTMLDivElement>,
+) => React.JSX.Element) & { propTypes?: any };
+
+const BrowserMultiInputDateRangeField = React.forwardRef(
+  (props: BrowserMultiInputDateRangeFieldProps, ref: React.Ref<HTMLDivElement>) => {
+    const {
+      slotProps,
+      value,
+      defaultValue,
+      format,
+      onChange,
+      readOnly,
+      disabled,
+      onError,
+      shouldDisableDate,
+      minDate,
+      maxDate,
+      disableFuture,
+      disablePast,
+      selectedSections,
+      onSelectedSectionsChange,
+      className,
+    } = props;
+
+    const { inputRef: startInputRef, ...startTextFieldProps } = useSlotProps({
+      elementType: 'input',
+      externalSlotProps: slotProps?.textField,
+      ownerState: { ...props, position: 'start' },
+    }) as MultiInputFieldSlotTextFieldProps;
+
+    const { inputRef: endInputRef, ...endTextFieldProps } = useSlotProps({
+      elementType: 'input',
+      externalSlotProps: slotProps?.textField,
+      ownerState: { ...props, position: 'end' },
+    }) as MultiInputFieldSlotTextFieldProps;
+
+    const {
+      startDate: { ref: startRef, ...startDateProps },
+      endDate: { ref: endRef, ...endDateProps },
+    } = useMultiInputDateRangeField<Dayjs, MultiInputFieldSlotTextFieldProps>({
+      sharedProps: {
+        value,
+        defaultValue,
+        format,
+        onChange,
+        readOnly,
+        disabled,
+        onError,
+        shouldDisableDate,
+        minDate,
+        maxDate,
+        disableFuture,
+        disablePast,
+        selectedSections,
+        onSelectedSectionsChange,
+      },
+      startTextFieldProps,
+      endTextFieldProps,
+      startInputRef,
+      endInputRef,
+    });
+
+    return (
+      <Stack ref={ref} spacing={2} direction="row" className={className}>
+        <BrowserField {...startDateProps} inputRef={startRef} />
+        <span> — </span>
+        <BrowserField {...endDateProps} inputRef={endRef} />
+      </Stack>
+    );
+  },
+) as BrowserMultiInputDateRangeFieldComponent;
+
+const BrowserDateRangePicker = React.forwardRef(
+  (props: DateRangePickerProps<Dayjs>, ref: React.Ref<HTMLDivElement>) => {
+    return (
+      <DateRangePicker
+        ref={ref}
+        {...props}
+        slots={{ ...props?.slots, field: BrowserMultiInputDateRangeField }}
+      />
+    );
+  },
+);
+
+export default function RangePickerWithBrowserField() {
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <BrowserDateRangePicker />
+    </LocalizationProvider>
+  );
+}
