+++ docs/data/date-pickers/custom-field/RangePickerWithBrowserField.tsx.preview
@@ -1,0 +1,1 @@
+<BrowserDateRangePicker />
