+++ docs/data/date-pickers/custom-field/RangePickerWithSingleInputBrowserField.tsx
@@ -1,0 +1,193 @@
+import * as React from 'react';
+import { Dayjs } from 'dayjs';
+import { unstable_useForkRef as useForkRef } from '@mui/utils';
+import { useSlotProps } from '@mui/base/utils';
+import Box from '@mui/material/Box';
+import IconButton from '@mui/material/IconButton';
+import InputAdornment from '@mui/material/InputAdornment';
+import { DateRangeIcon } from '@mui/x-date-pickers/icons';
+import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
+import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
+import {
+  DateRangePicker,
+  DateRangePickerProps,
+} from '@mui/x-date-pickers-pro/DateRangePicker';
+import {
+  unstable_useSingleInputDateRangeField as useSingleInputDateRangeField,
+  SingleInputDateRangeFieldProps,
+} from '@mui/x-date-pickers-pro/SingleInputDateRangeField';
+import { useClearableField } from '@mui/x-date-pickers/hooks';
+import type {
+  SingleInputDateRangeFieldSlotsComponent,
+  SingleInputDateRangeFieldSlotsComponentsProps,
+} from '@mui/x-date-pickers-pro/SingleInputDateRangeField/SingleInputDateRangeField.types';
+
+interface BrowserFieldProps
+  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'> {
+  label?: React.ReactNode;
+  inputRef?: React.Ref<any>;
+  InputProps?: {
+    ref?: React.Ref<any>;
+    endAdornment?: React.ReactNode;
+    startAdornment?: React.ReactNode;
+  };
+  error?: boolean;
+  focused?: boolean;
+  ownerState?: any;
+  sx?: any;
+}
+
+type BrowserFieldComponent = ((
+  props: BrowserFieldProps & React.RefAttributes<HTMLDivElement>,
+) => React.JSX.Element) & { propTypes?: any };
+
+const BrowserField = React.forwardRef(
+  (props: BrowserFieldProps, ref: React.Ref<HTMLDivElement>) => {
+    const {
+      disabled,
+      id,
+      label,
+      inputRef,
+      InputProps: { ref: containerRef, startAdornment, endAdornment } = {},
+      // extracting `error`, 'focused', and `ownerState` as `input` does not support those props
+      error,
+      focused,
+      ownerState,
+      sx,
+      ...other
+    } = props;
+
+    const handleRef = useForkRef(containerRef, ref);
+
+    return (
+      <Box
+        sx={{ ...(sx || {}), display: 'flex', alignItems: 'center' }}
+        id={id}
+        ref={handleRef}
+      >
+        {startAdornment}
+        <input disabled={disabled} ref={inputRef} {...other} />
+        {endAdornment}
+      </Box>
+    );
+  },
+) as BrowserFieldComponent;
+
+interface BrowserSingleInputDateRangeFieldProps
+  extends SingleInputDateRangeFieldProps<
+    Dayjs,
+    Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'>
+  > {
+  onAdornmentClick?: () => void;
+}
+
+type BrowserSingleInputDateRangeFieldComponent = ((
+  props: BrowserSingleInputDateRangeFieldProps & React.RefAttributes<HTMLDivElement>,
+) => React.JSX.Element) & { fieldType?: string };
+
+const BrowserSingleInputDateRangeField = React.forwardRef(
+  (props: BrowserSingleInputDateRangeFieldProps, ref: React.Ref<HTMLDivElement>) => {
+    const { slots, slotProps, onAdornmentClick, ...other } = props;
+
+    const {
+      inputRef: externalInputRef,
+      ...textFieldProps
+    }: SingleInputDateRangeFieldProps<Dayjs> = useSlotProps({
+      elementType: 'input',
+      externalSlotProps: slotProps?.textField,
+      externalForwardedProps: other,
+      ownerState: props as any,
+    });
+
+    const {
+      ref: inputRef,
+      onClear,
+      clearable,
+      ...fieldProps
+    } = useSingleInputDateRangeField<Dayjs, typeof textFieldProps>({
+      props: textFieldProps,
+      inputRef: externalInputRef,
+    });
+
+    /* If you don't need a clear button, you can skip the use of this hook */
+    const { InputProps: ProcessedInputProps, fieldProps: processedFieldProps } =
+      useClearableField<
+        {},
+        typeof textFieldProps.InputProps,
+        SingleInputDateRangeFieldSlotsComponent,
+        SingleInputDateRangeFieldSlotsComponentsProps<Dayjs>
+      >({
+        onClear,
+        clearable,
+        fieldProps,
+        InputProps: {
+          ...fieldProps.InputProps,
+          endAdornment: (
+            <InputAdornment position="end">
+              <IconButton onClick={onAdornmentClick}>
+                <DateRangeIcon />
+              </IconButton>
+            </InputAdornment>
+          ),
+        },
+        slots,
+        slotProps,
+      });
+
+    return (
+      <BrowserField
+        {...processedFieldProps}
+        ref={ref}
+        style={{
+          minWidth: 300,
+        }}
+        inputRef={inputRef}
+        InputProps={{ ...ProcessedInputProps }}
+      />
+    );
+  },
+) as BrowserSingleInputDateRangeFieldComponent;
+
+BrowserSingleInputDateRangeField.fieldType = 'single-input';
+
+const BrowserSingleInputDateRangePicker = React.forwardRef(
+  (props: DateRangePickerProps<Dayjs>, ref: React.Ref<HTMLDivElement>) => {
+    const [isOpen, setIsOpen] = React.useState(false);
+
+    const toggleOpen = () => setIsOpen((currentOpen) => !currentOpen);
+
+    const handleOpen = () => setIsOpen(true);
+
+    const handleClose = () => setIsOpen(false);
+
+    return (
+      <DateRangePicker
+        ref={ref}
+        {...props}
+        open={isOpen}
+        onClose={handleClose}
+        onOpen={handleOpen}
+        slots={{ field: BrowserSingleInputDateRangeField }}
+        slotProps={{
+          ...props?.slotProps,
+          field: {
+            onAdornmentClick: toggleOpen,
+            ...props?.slotProps?.field,
+          } as any,
+        }}
+      />
+    );
+  },
+);
+
+export default function RangePickerWithSingleInputBrowserField() {
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <BrowserSingleInputDateRangePicker
+        slotProps={{
+          field: { clearable: true },
+        }}
+      />
+    </LocalizationProvider>
+  );
+}
