+++ docs/data/date-pickers/custom-field/RangePickerWithSingleInputBrowserField.tsx.preview
@@ -1,4 +1,4 @@
-<BrowserDatePicker
+<BrowserSingleInputDateRangePicker
   slotProps={{
     field: { clearable: true },
   }}
